--- a/aio/scripts/switch-to-viewengine.js
+++ b/aio/scripts/switch-to-viewengine.js
@@ -4,16 +4,13 @@
 const {extend, parse} = require('cjson');
 const {readFileSync, writeFileSync} = require('fs');
 const {join, resolve} = require('path');
-const {exec, set} = require('shelljs');
-
-set('-e');
 
 // Constants
 const ROOT_DIR = resolve(__dirname, '..');
 const NG_JSON = join(ROOT_DIR, 'angular.json');
 const NG_COMPILER_OPTS = {
   angularCompilerOptions: {
-    enableIvy: true,
+    enableIvy: false,
   },
 };
 
@@ -26,7 +23,7 @@ function _main() {
   const ngConfig = parse(readFileSync(NG_JSON, 'utf8'));
   const tsConfigPath = join(ROOT_DIR, ngConfig.projects.site.architect.build.options.tsConfig);
 
-  // Enable Ivy in TS config.
+  // Enable ViewIngine/Disable Ivy in TS config.
   console.log(`\nModifying \`${tsConfigPath}\`...`);
   const oldTsConfigStr = readFileSync(tsConfigPath, 'utf8');
   const oldTsConfigObj = parse(oldTsConfigStr);
@@ -35,13 +32,8 @@ function _main() {
   console.log(`\nNew config: ${newTsConfigStr}`);
   writeFileSync(tsConfigPath, newTsConfigStr);
 
-  // Run ngcc.
-  const ngccArgs = '--loglevel debug --properties es2015';
-  console.log(`\nRunning ngcc (with args: ${ngccArgs})...`);
-  exec(`yarn ivy-ngcc ${ngccArgs}`);
-
   // Done.
-  console.log('\nReady to build with Ivy!');
-  console.log('(To switch back to ViewEngine (with packages from npm), undo the changes in ' +
+  console.log('\nReady to build with ViewEngine!');
+  console.log('(To switch back to Ivy (with packages from npm), undo the changes in ' +
               `\`${tsConfigPath}\` and run \`yarn aio-use-npm && yarn example-use-npm\`.)`);
 }