--- a/aio/tsconfig.json
+++ b/aio/tsconfig.json
@@ -33,7 +33,4 @@
     "scripts",
     "tools"
   ],
-  "angularCompilerOptions": {
-    "enableIvy": false,
-  },
 }