--- a/aio/tsconfig.json
+++ b/aio/tsconfig.json
@@ -7,7 +7,6 @@
     "declaration": false,
     "module": "esnext",
     "moduleResolution": "node",
-    "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "importHelpers": true,
     "target": "es2015",
@@ -33,4 +32,10 @@
     "scripts",
     "tools"
   ],
+  "angularCompilerOptions": {
+    "disableTypeScriptVersionCheck": true,
+    // TODO: Fix template type errors and enable.
+    // "fullTemplateTypeCheck": true,
+    "strictInjectionParameters": true
+  }
 }