--- a/packages/bazel/src/ng_package/ng_package.bzl
+++ b/packages/bazel/src/ng_package/ng_package.bzl
@@ -36,7 +36,7 @@ load("//packages/bazel/src/ng_package:collect-type-definitions.bzl", "collect_ty
 # Prints a debug message if "--define=VERBOSE_LOGS=true" is specified.
 def _debug(vars, *args):
     if "VERBOSE_LOGS" in vars.keys():
-        print("[ng_package.bzl]", *args)
+        print("[ng_package.bzl]", args)
 
 _DEFAULT_NG_PACKAGER = "@npm//@angular/bazel/bin:packager"