The angular.io project uses Angular and CLI v9, which by default turns on Ivy mode. However, since ec4381d, we explicitly opt out of Ivy.

This commit removes the `enabledIvy: false` configuration, thus allowing the default behavior of having Ivy on.

NOTE:
This commit only changes the angular.io projects. The docs examples need to be updated separately (first to Angular and CLI v9 and then to Ivy).

##
Jira issue: [FW-413](https://angular-team.atlassian.net/browse/FW-413)
(I also created [FW-1607](https://angular-team.atlassian.net/browse/FW-1607) to track updating the docs examples to latest `@angular/*` and Ivy.)

---
I haven't profiled this, but the performance feels on par with ViewEngine when using the app.

There is currently a difference in payload sizes however. The lazy-loaded bundles are a little smaller with Ivy, but the main (eagerly loaded) bundle (and thus the overall app size) has increased (see below for details).

> _EDIT: Updating to CLI v9.0.0-next.6 improves things a lot for Ivy (bringing the different to ~2%, but there also seems to be a ~12KB regression in in the ES2015 main bundle for ViewEngine._ 🤔 

<details>
  <summary><b>Sizes with cli@9.0.0-next.4</b> (~10% increase with Ivy)</summary>

<p align="right"><sub><i>* Bold numbers indicate an undesirable change.</i></sub></p>

**ES2015**

| File | VE (B) | IVY (B) | Diff (B) | Diff (%) |
| --- | ---: | ---: | ---: | ---: |
| **_Eager bundles_** |
| `main` | 436411 | 496036 | **59625** | **14%** |
| `polyfills` | 52503 | 52503 | _0_ | _0%_ |
| `runtime` | 3097 | 2987 | _-110_ | _-4%_ |
| **_Total_** | 492011 | 551526 | **59515** | **12%** |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_Lazy bundles_** |
| `announcement-bar-announcement-bar-module` | 2632 | 1848 | _-784_ | _-30%_ |
| `api-api-list-module` | 8349 | 6261 | _-2088_ | _-25%_ |
| `code-code-example-module` | 3120 | 2237 | _-883_ | _-28%_ |
| `code-code-tabs-module` | 35930 | 33545 | _-2385_ | _-7%_ |
| `contributor-contributor-list-module` | 7417 | 5327 | _-2090_ | _-28%_ |
| `default~code-code-example-module~code-code-tabs-module` | 52597 | 53171 | **574** | **1%** |
| `live-example-live-example-module` | 5239 | 4270 | _-969_ | _-18%_ |
| `resource-resource-list-module` | 4702 | 3460 | _-1242_ | _-26%_ |
| `search-file-not-found-search-module` | 1567 | 1213 | _-354_ | _-23%_ |
| `toc-toc-module` | 6373 | 5209 | _-1164_ | _-18%_ |
| **_Total_** | 127926 | 116541 | _-11385_ | _-9%_ |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_TOTAL_** | 619937 | 668067 | **48130** | **8%** |

---
<p align="right"><sub><i>* Bold numbers indicate an undesirable change.</i></sub></p>

**ES5**

| File | VE (B) | IVY (B) | Diff (B) | Diff (%) |
| --- | ---: | ---: | ---: | ---: |
| **_Eager bundles_** |
| `main` | 481663 | 541663 | **60000** | **12%** |
| `polyfills` | 142239 | 142239 | _0_ | _0%_ |
| `runtime` | 3091 | 2981 | _-110_ | _-4%_ |
| **_Total_** | 626993 | 686883 | **59890** | **10%** |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_Lazy bundles_** |
| `announcement-bar-announcement-bar-module` | 2767 | 1954 | _-813_ | _-29%_ |
| `api-api-list-module` | 9084 | 6942 | _-2142_ | _-24%_ |
| `code-code-example-module` | 3536 | 2614 | _-922_ | _-26%_ |
| `code-code-tabs-module` | 38140 | 35244 | _-2896_ | _-8%_ |
| `contributor-contributor-list-module` | 7616 | 5431 | _-2185_ | _-29%_ |
| `default~code-code-example-module~code-code-tabs-module` | 57173 | 57641 | **468** | **1%** |
| `live-example-live-example-module` | 5548 | 4525 | _-1023_ | _-18%_ |
| `resource-resource-list-module` | 4838 | 3548 | _-1290_ | _-27%_ |
| `search-file-not-found-search-module` | 1603 | 1218 | _-385_ | _-24%_ |
| `toc-toc-module` | 6778 | 5567 | _-1211_ | _-18%_ |
| **_Total_** | 137083 | 124684 | _-12399_ | _-9%_ |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_TOTAL_** | 764076 | 811567 | **47491** | **6%** |

</details>

---
<details>
  <summary><b>Sizes with cli@9.0.0-next.6</b> (~2% increase with Ivy)</summary>

<p align="right"><sub><i>* Bold numbers indicate an undesirable change.</i></sub></p>

**ES2015**

| File | VE (B) | IVY (B) | Diff (B) | Diff (%) |
| --- | ---: | ---: | ---: | ---: |
| **_Eager bundles_** |
| `main` | 448708 | 458422 | **9714** | **2%** |
| `polyfills` | 52608 | 52608 | _0_ | _0%_ |
| `runtime` | 3106 | 2996 | _-110_ | _-4%_ |
| **_Total_** | 504422 | 514026 | **9604** | **2%** |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_Lazy bundles_** |
| `announcement-bar-announcement-bar-module` | 2632 | 1848 | _-784_ | _-30%_ |
| `api-api-list-module` | 8357 | 6265 | _-2092_ | _-25%_ |
| `code-code-example-module` | 3120 | 2237 | _-883_ | _-28%_ |
| `code-code-tabs-module` | 35998 | 33582 | _-2416_ | _-7%_ |
| `contributor-contributor-list-module` | 7445 | 5341 | _-2104_ | _-28%_ |
| `default~code-code-example-module~code-code-tabs-module` | 52860 | 53411 | **551** | **1%** |
| `live-example-live-example-module` | 5247 | 4270 | _-977_ | _-19%_ |
| `resource-resource-list-module` | 4702 | 3460 | _-1242_ | _-26%_ |
| `search-file-not-found-search-module` | 1567 | 1213 | _-354_ | _-23%_ |
| `toc-toc-module` | 6395 | 5223 | _-1172_ | _-18%_ |
| **_Total_** | 128323 | 116850 | _-11473_ | _-9%_ |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_TOTAL_** | 632745 | 630876 | _-1869_ | _0%_ |

---
<p align="right"><sub><i>* Bold numbers indicate an undesirable change.</i></sub></p>

**ES5**

| File | VE (B) | IVY (B) | Diff (B) | Diff (%) |
| --- | ---: | ---: | ---: | ---: |
| **_Eager bundles_** |
| `main` | 482744 | 501393 | **18649** | **4%** |
| `polyfills` | 142383 | 142383 | _0_ | _0%_ |
| `runtime` | 3100 | 2990 | _-110_ | _-4%_ |
| **_Total_** | 628227 | 646766 | **18539** | **3%** |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_Lazy bundles_** |
| `announcement-bar-announcement-bar-module` | 2767 | 1954 | _-813_ | _-29%_ |
| `api-api-list-module` | 9092 | 6946 | _-2146_ | _-24%_ |
| `code-code-example-module` | 3536 | 2614 | _-922_ | _-26%_ |
| `code-code-tabs-module` | 38206 | 35279 | _-2927_ | _-8%_ |
| `contributor-contributor-list-module` | 7644 | 5443 | _-2201_ | _-29%_ |
| `default~code-code-example-module~code-code-tabs-module` | 57399 | 57847 | **448** | **1%** |
| `live-example-live-example-module` | 5556 | 4525 | _-1031_ | _-19%_ |
| `resource-resource-list-module` | 4838 | 3548 | _-1290_ | _-27%_ |
| `search-file-not-found-search-module` | 1603 | 1218 | _-385_ | _-24%_ |
| `toc-toc-module` | 6798 | 5579 | _-1219_ | _-18%_ |
| **_Total_** | 137439 | 124953 | _-12486_ | _-9%_ |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_TOTAL_** | 765666 | 771719 | **6053** | **1%** |

</details>

---
<details open>
  <summary><b>Sizes with cli@9.0.0-next.6 <a href="https://github.com/angular/angular/pull/32923#issuecomment-537370090">with terser@4.3.2</a></b> (~3% increase with Ivy)</summary>

<p align="right"><sub><i>* Bold numbers indicate an undesirable change.</i></sub></p>

**ES2015**

| File | VE (B) | IVY (B) | Diff (B) | Diff (%) |
| --- | ---: | ---: | ---: | ---: |
| **_Eager bundles_** |
| `main` | 437453 | 456861 | **19408** | **4%** |
| `polyfills` | 52608 | 52608 | _0_ | _0%_ |
| `runtime` | 3106 | 2996 | _-110_ | _-4%_ |
| **_Total_** | 493167 | 512465 | **19298** | **4%** |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_Lazy bundles_** |
| `announcement-bar-announcement-bar-module` | 2632 | 1848 | _-784_ | _-30%_ |
| `api-api-list-module` | 8357 | 6265 | _-2092_ | _-25%_ |
| `code-code-example-module` | 3120 | 2237 | _-883_ | _-28%_ |
| `code-code-tabs-module` | 35998 | 33582 | _-2416_ | _-7%_ |
| `contributor-contributor-list-module` | 7445 | 5341 | _-2104_ | _-28%_ |
| `default~code-code-example-module~code-code-tabs-module` | 52860 | 53411 | **551** | **1%** |
| `live-example-live-example-module` | 5247 | 4270 | _-977_ | _-19%_ |
| `resource-resource-list-module` | 4702 | 3460 | _-1242_ | _-26%_ |
| `search-file-not-found-search-module` | 1567 | 1213 | _-354_ | _-23%_ |
| `toc-toc-module` | 6395 | 5223 | _-1172_ | _-18%_ |
| **_Total_** | 128323 | 116850 | _-11473_ | _-9%_ |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_TOTAL_** | 621490 | 629315 | **7825** | **1%** |

---
<p align="right"><sub><i>* Bold numbers indicate an undesirable change.</i></sub></p>

**ES5**

| File | VE (B) | IVY (B) | Diff (B) | Diff (%) |
| --- | ---: | ---: | ---: | ---: |
| **_Eager bundles_** |
| `main` | 482632 | 501361 | **18729** | **4%** |
| `polyfills` | 142383 | 142383 | _0_ | _0%_ |
| `runtime` | 3100 | 2990 | _-110_ | _-4%_ |
| **_Total_** | 628115 | 646734 | **18619** | **3%** |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_Lazy bundles_** |
| `announcement-bar-announcement-bar-module` | 2767 | 1954 | _-813_ | _-29%_ |
| `api-api-list-module` | 9092 | 6946 | _-2146_ | _-24%_ |
| `code-code-example-module` | 3536 | 2614 | _-922_ | _-26%_ |
| `code-code-tabs-module` | 38206 | 35279 | _-2927_ | _-8%_ |
| `contributor-contributor-list-module` | 7644 | 5443 | _-2201_ | _-29%_ |
| `default~code-code-example-module~code-code-tabs-module` | 57399 | 57847 | **448** | **1%** |
| `live-example-live-example-module` | 5556 | 4525 | _-1031_ | _-19%_ |
| `resource-resource-list-module` | 4838 | 3548 | _-1290_ | _-27%_ |
| `search-file-not-found-search-module` | 1603 | 1218 | _-385_ | _-24%_ |
| `toc-toc-module` | 6798 | 5579 | _-1219_ | _-18%_ |
| **_Total_** | 137439 | 124953 | _-12486_ | _-9%_ |
| &nbsp; | &nbsp; | &nbsp; | &nbsp; | &nbsp; |
| **_TOTAL_** | 765554 | 771687 | **6133** | **1%** |

</details>