@filipesilva 2019-10-02
Size difference might be related to https://github.com/terser/terser/issues/470. `terser@4.3.4` is transitively used, or at least it appears in the yarn lock file. You can try and set the yarn resolution to `4.3.2` for now. The upcoming `4.3.5` release of terser should fix the regression.