@gkalpak 2019-10-02
I pinned terser to 4.3.2, as suggested by @filipesilva, and the 12KB ViewEngine regression seems to go away. I updated the numbers.