@gkalpak 2019-10-02
I just noticed a bug:
The main content area does not shrink to leave room for the ToC on the right (thus covering it up and essentially hiding it entirely). Investigating...