@gkalpak 2019-10-02
Updating Angular to 9.0.0-next.9 fixed the ToC issue :tada:

I have added 3 new commits (since last review), with the following changes:
- ``refactor(docs-infra): replace `System.import()` with `import()` `` (to avoid warning)
- ``build(docs-infra): align config with what cli generates for new apps`` (to avoid more warnings)
- ``build(docs-infra): update @angular/* to 9.0.0-next.9`` (to fix ToC bug).

PTAL